เพิ่ม confirm box ลบรายการยาหน่วยเบิก แก้รายการยาถึงจุดเบิก expannd ไ…

+++ src/app/staff/main/main.component.ts
@@ -203,14 +203,18 @@ export class MainComponent implements OnInit {
   async deleteGeneric(genericId) {
     try {
       this.modalLoading.show();
-      const rs = await this.staffService.removeGeneric(genericId);
+      this.alertService.confirm('คุณต้องการลบรายการนี้')
+        .then(async () => {
+          const rs = await this.staffService.removeGeneric(genericId);
+          this.modalLoading.hide();
+          if (rs.ok) {
+            await this.getGenerics();
+            this.alertService.success();
+          } else {
+            this.alertService.error(rs.error)
+          }
+        })
       this.modalLoading.hide();
-      if (rs.ok) {
-        await this.getGenerics();
-        this.alertService.success();
-      } else {
-        this.alertService.error(rs.error)
-      }
     } catch (error) {
       this.modalLoading.hide();
       this.alertService.error(error);
@@ -220,14 +224,18 @@ export class MainComponent implements OnInit {
   async deleteProduct(productId) {
     try {
       this.modalLoading.show();
-      const rs = await this.staffService.removeProduct(productId);
+      this.alertService.confirm('คุณต้องการลบรายการนี้')
+        .then(async () => {
+          const rs = await this.staffService.removeProduct(productId);
+          this.modalLoading.hide();
+          if (rs.ok) {
+            await this.getProducts();
+            this.alertService.success();
+          } else {
+            this.alertService.error(rs.error)
+          }
+        })
       this.modalLoading.hide();
-      if (rs.ok) {
-        await this.getProducts();
-        this.alertService.success();
-      } else {
-        this.alertService.error(rs.error)
-      }
     } catch (error) {
       this.modalLoading.hide();
       this.alertService.error(error);
